more refactoring done, finished with handlers
