+++ server/src/main/java/handler/GamesList.java
@@ -1,0 +1,13 @@
+package handler;
+
+import model.GameData;
+
+import java.util.ArrayList;
+
+public class GamesList {
+    private ArrayList<GameData> games;
+
+    public GamesList(ArrayList<GameData> games) {
+        this.games = games;
+    }
+}
