+++ server/src/main/java/handler/LoginHandler.java
@@ -7,10 +7,8 @@ import spark.Request;
 import spark.Response;
 import spark.Route;
 
-import static handler.Handler.GSON;
-import static handler.Handler.userService;
 
-public class LoginHandler implements Route {
+public class LoginHandler extends Handler implements Route {
     public Object handle(Request request, Response response) throws DataAccessException {
         UserData userData = GSON.fromJson(request.body(), UserData.class);
         AuthData authData = userService.login(userData);
