+++ server/src/main/java/handler/RegisterHandler.java
@@ -7,10 +7,7 @@ import spark.Request;
 import spark.Response;
 import spark.Route;
 
-import static handler.Handler.GSON;
-import static handler.Handler.userService;
-
-public class RegisterHandler implements Route {
+public class RegisterHandler extends Handler implements Route {
     public Object handle(Request request, Response response) throws DataAccessException {
         UserData userData = GSON.fromJson(request.body(), UserData.class);
         AuthData authData = userService.register(userData);
